# apps/backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "VRK API",
        "contact": {}
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/employee-invites": {
            "get": {
                "description": "Returns the employee invite lifecycle visible to the active organization admin.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee-invites"
                ],
                "summary": "List employee invites",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/bootstrap.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/bootstrap.EmployeeInviteResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/bootstrap.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/bootstrap.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a draft employee invite for the active organization admin with role template, scope, and expiry policy.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee-invites"
                ],
                "summary": "Create employee invite draft",
                "parameters": [
                    {
                        "description": "Employee invite payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bootstrap.CreateEmployeeInviteRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/bootstrap.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/bootstrap.EmployeeInviteResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/bootstrap.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/bootstrap.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/bootstrap.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/bootstrap.Response"
                        }
                    }
                }
            }
        },
        "/employee-invites/{inviteID}/revoke": {
            "post": {
                "description": "Revokes a draft, sent, or opened employee invite.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee-invites"
                ],
                "summary": "Revoke employee invite",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Employee invite ID",
                        "name": "inviteID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/bootstrap.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/bootstrap.EmployeeInviteResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/bootstrap.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/bootstrap.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/bootstrap.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/bootstrap.Response"
                        }
                    }
                }
            }
        },
        "/employee-invites/{inviteID}/send": {
            "post": {
                "description": "Transitions a draft employee invite to sent and issues a one-time token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee-invites"
                ],
                "summary": "Send employee invite",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Employee invite ID",
                        "name": "inviteID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/bootstrap.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/bootstrap.EmployeeInviteResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/bootstrap.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/bootstrap.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/bootstrap.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/bootstrap.Response"
                        }
                    }
                }
            }
        },
...
        "/first-admin-invites/{token}": {
            "get": {
                "description": "Opens the first-admin invite link before password setup.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bootstrap"
                ],
                "summary": "Inspect first-admin invite",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invite token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/bootstrap.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/bootstrap.InviteInspectionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/bootstrap.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/bootstrap.Response"
                        }
                    }
                }
            }
        },
        "/first-admin-invites/{token}/accept": {
            "post": {
                "description": "Sets the invited admin password, creates membership and organization-admin grant, and starts a session.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bootstrap"
                ],
                "summary": "Accept first-admin invite",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invite token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Password payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bootstrap.AcceptInviteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/bootstrap.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/bootstrap.SessionSummaryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/bootstrap.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/bootstrap.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/bootstrap.Response"
                        }
                    }
                }
            }
        },
        "/invites/{token}": {
            "get": {
                "description": "Opens a public invite link and returns the current invite state for first-admin and employee activation flows.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bootstrap"
                ],
                "summary": "Inspect public invite",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invite token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/bootstrap.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/bootstrap.PublicInviteInspectionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/bootstrap.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/bootstrap.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "bootstrap.AcceptInviteRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                }
            }
        },
...
        "bootstrap.CreateEmployeeInviteRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "expiresAt": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "roleTemplate": {
                    "type": "string"
                },
                "scopeId": {
                    "type": "string"
                },
                "scopeType": {
                    "type": "string"
                }
            }
        },
...
        "bootstrap.EmployeeInviteResponse": {
            "type": "object",
            "properties": {
                "acceptPath": {
                    "type": "string"
                },
                "acceptedAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "expiresAt": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "inviteToken": {
                    "type": "string"
                },
                "openedAt": {
                    "type": "string"
                },
                "revokedAt": {
                    "type": "string"
                },
                "roleTemplate": {
                    "type": "string"
                },
                "scopeId": {
                    "type": "string"
                },
                "scopeLabel": {
                    "type": "string"
                },
                "scopeType": {
                    "type": "string"
                },
                "sentAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "bootstrap.InviteInspectionResponse": {
            "type": "object",
            "properties": {
                "firstAdminName": {
                    "type": "string"
                },
                "inviteEmail": {
                    "type": "string"
                },
                "inviteExpiresAt": {
                    "type": "string"
                },
                "inviteStatus": {
                    "type": "string"
                },
                "launchState": {
                    "type": "string"
                },
                "organizationId": {
                    "type": "string"
                },
                "organizationName": {
                    "type": "string"
                },
                "organizationRole": {
                    "type": "string"
                }
            }
        },
...
        "bootstrap.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}